Clean up code to code quality standards Change package name from Jerry to jerry, some variables to final, and remove commented code
