+++ src/main/java/jerry/Storage.java
@@ -1,4 +1,4 @@
-package Jerry;
+package jerry;
 
 import java.io.File;
 import java.io.FileNotFoundException;
