+++ src/main/java/jerry/command/CommandFormatException.java
@@ -1,4 +1,4 @@
-package Jerry.command;
+package jerry.command;
 
 public class CommandFormatException extends Exception {
     public CommandFormatException(String message) {
