+++ src/main/java/jerry/command/InvalidCommand.java
@@ -1,6 +1,6 @@
-package Jerry.command;
+package jerry.command;
 
-import Jerry.Ui;
+import jerry.Ui;
 
 public class InvalidCommand extends Command{
