+++ src/test/java/jerry/StorageTest.java
@@ -1,4 +1,4 @@
-package Jerry;
+package jerry;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
