+++ src/test/java/jerry/TaskListTest.java
@@ -1,4 +1,4 @@
-package Jerry;
+package jerry;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
