+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Jerry.
+Hello! I'm jerry.
 Anything I can do for you?
 Got it. I've added this task:
   [T][ ] read book
